Honor discoverable field
Added a few changes to honor the discoverable field in Mastodon accounts:

- `get_compact_data` is moved to `mastodon.py` and upated to skip those
  statuses where `account.discoverable` is False
- `notebook.py` is updated to look for the method inside `mastodon.py`
- `demo.py` is updated to remove the method (which was not used anyway,
  as we are not using paginated data but directly loading dataframes)
